fix(api): múltiplo de venda decimal derrubava o lote de produtos do Dlinks
coalesce($6, 1) fazia o Postgres inferir o parâmetro como integer;
um produto com multiplo_venda 10.1 respondia 500 ("sintaxe inválida
para tipo integer") e o restante do lote não era gravado. A coluna
qtd_por_embalagem é numeric(10,3), então basta o cast explícito.

+++ apps/api/src/integracoes-dlinks/dlinks-sync.service.ts
@@ -88,7 +88,7 @@ export class DlinksSyncService {
            $1, $1, $2,
            (select id from marcas where erp_marca_id = $3),
            (select id from categorias where erp_categoria_id = $4),
-           $5, coalesce($6, 1), $1, now()
+           $5, coalesce($6::numeric, 1), $1, now()
          )
          on conflict (erp_produto_id) do update set
            nome = excluded.nome,
